Move helpers into directives
