+++ packages/lit-html/package.json
@@ -53,6 +53,14 @@
       "development": "./development/directives/if-defined.js",
       "default": "./directives/if-defined.js"
     },
+    "./directives/join.js": {
+      "development": "./development/directives/join.js",
+      "default": "./directives/join.js"
+    },
+    "./directives/map.js": {
+      "development": "./development/directives/map.js",
+      "default": "./directives/map.js"
+    },
     "./directives/live.js": {
       "development": "./development/directives/live.js",
       "default": "./directives/live.js"
@@ -60,6 +68,10 @@
     "./directives/ref.js": {
       "development": "./development/directives/ref.js",
       "default": "./directives/ref.js"
+    },
+    "./directives/range.js": {
+      "development": "./development/directives/range.js",
+      "default": "./directives/range.js"
     },
     "./directives/repeat.js": {
       "development": "./development/directives/repeat.js",
@@ -85,13 +97,13 @@
       "development": "./development/directives/until.js",
       "default": "./directives/until.js"
     },
+    "./directives/when.js": {
+      "development": "./development/directives/when.js",
+      "default": "./directives/when.js"
+    },
     "./experimental-hydrate.js": {
       "development": "./development/experimental-hydrate.js",
       "default": "./experimental-hydrate.js"
-    },
-    "./helpers.js": {
-      "development": "./development/helpers.js",
-      "default": "./helpers.js"
     },
     "./polyfill-support.js": {
       "development": "./development/polyfill-support.js",
@@ -114,11 +126,11 @@
     "build:version-stability-test": "rollup -c rollup-version-stability-test.config.js",
     "check-version": "node scripts/check-version-tracker.js",
     "checksize": "rollup -c --environment=CHECKSIZE",
-    "clean": "rm -rf {async-directive,directive-helpers,directive,experimental-hydrate,helpers,lit-html,polyfill-support,private-ssr-support,static}.{js,js.map,d.ts} test/ directives/ development/ version-stability-test/ *.tsbuildinfo",
+    "clean": "rm -rf {async-directive,directive-helpers,directive,experimental-hydrate,lit-html,polyfill-support,private-ssr-support,static}.{js,js.map,d.ts} test/ directives/ development/ version-stability-test/ *.tsbuildinfo",
     "dev": "scripts/dev.sh",
     "prepublishOnly": "npm run check-version",
     "test": "npm run test:dev && npm run test:prod",
-    "test:dev": "cd ../tests && npx wtr '../lit-html/development/**/helpers_test.(js|html)'",
+    "test:dev": "cd ../tests && npx wtr '../lit-html/development/**/*_test.(js|html)'",
     "test:prod": "MODE=prod npm run test:dev -- '../lit-html/src/test/version-stability_test.js'",
     "test:watch": "npm run test:dev -- --watch"
   },
@@ -127,7 +139,6 @@
     "/directive-helpers.{d.ts,d.ts.map,js,js.map}",
     "/directive.{d.ts,d.ts.map,js,js.map}",
     "/experimental-hydrate.{d.ts,d.ts.map,js,js.map}",
-    "/helpers.{d.ts,d.ts.map,js,js.map}",
     "/lit-html.{d.ts,d.ts.map,js,js.map}",
     "/polyfill-support.{d.ts,d.ts.map,js,js.map}",
     "/private-ssr-support.{d.ts,d.ts.map,js,js.map}",
