+++ packages/lit/src/directives/join.ts
@@ -1,0 +1,7 @@
+/**
+ * @license
+ * Copyright 2021 Google LLC
+ * SPDX-License-Identifier: BSD-3-Clause
+ */
+
+export * from 'lit-html/directives/join.js';
